+++ src/v1/components/counters/counterErrors.js
@@ -1,0 +1,6 @@
+const ErrorHandler = require("../../classes/ErrorHandler");
+
+
+module.exports = {
+  ErrorHandler
+}
