+++ src/v1/components/counters/index.js
@@ -1,0 +1,4 @@
+const route = require('./countersRouters');
+
+
+module.exports = route
